Hide inactive spaces with no open windows
Uses the pending chunkwm patch `tiling::query --windows-for-desktop`
to identify spaces without open windows and hide them from the pager.
Communicates with chunkwm directly via socket using a direct Swift
port of `chunkc`. This could be extended for additional functionality.

Socket polling on the main thread is a bad idea so the menu button
regeneration has been moved to a background thread, with the actual
assignment of the new image on the main thread.

Eventually this might need to be placed under a menu option, but
for my use case I don't want to see empty spaces so have made it the
default.